api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
psycopg2-binary==2.9.9
redis==5.0.1
minio==7.2.0
qdrant-client==1.7.0
sentence-transformers==2.2.2
torch==2.0.1
transformers==4.30.0
huggingface-hub==0.16.4
numpy==1.24.0
pydantic==2.5.0
python-dotenv==1.0.0
